% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\alias{add_weights}
\alias{add_residuals}
\alias{add_rstandard}
\alias{add_i2}
\alias{add_h2}
\alias{add_tau2}
\alias{add_blups}
\title{Add model values to a table of meta-analysis results}
\usage{
add_weights(x, .meta = NULL, group = NULL)

add_residuals(x, .meta = NULL)

add_rstandard(x, .meta = NULL)

add_i2(x, .meta = NULL)

add_h2(x, .meta = NULL)

add_tau2(x, .meta = NULL)

add_blups(x, .meta = NULL, exponentiate = FALSE)
}
\arguments{
\item{x}{a tidied meta-analysis}

\item{.meta}{a meta-analysis object. If not \code{NULL}, then looks for a \code{meta}
column in \code{x} containing an overall meta-analysis}

\item{group}{a grouping variable}

\item{exponentiate}{logical. Should the results be exponentiated?}
}
\value{
a \code{tbl}
}
\description{
\code{add_weights()} adds study weights to the model. \code{add_residuals()} and
\code{add_rstandard()} add residuals. \code{add_i2()}, \code{add_h2()}, \code{add_tau2()}
}
\examples{

library(dplyr)

iud_cxca \%>\%
  group_by(group) \%>\%
  meta_analysis(, yi = lnes, sei = selnes, slab = study_name) \%>\%
  add_i2()

}
