% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{add_meta}
\alias{add_meta}
\title{Add a meta-analysis to a tbl}
\usage{
add_meta(x, meta, ...)
}
\arguments{
\item{x}{a \code{data.frame} produced by \code{tidy()}}

\item{meta}{a meta-analysis object (e.g. \code{rma.uni})}

\item{...}{additional arguments to \code{as.tbl}}
}
\value{
a \code{tbl} with a \code{listcol}, \code{meta}, containing the meta-analysis
}
\description{
\code{add_meta()} add a meta-analysis object to a tidied meta-analysis.
}
\examples{

library(metafor)
library(broom)

ma <- rma(yi = lnes, sei = selnes, slab = study_name, data = iud_cxca)

ma \%>\%
  tidy() \%>\%
  add_meta(ma)

}
