% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.rma.uni}
\alias{augment.rma.uni}
\title{Augment data with values from a meta-analysis model}
\usage{
augment.rma.uni(x, ...)
}
\arguments{
\item{x}{a meta-analysis object. Currently supports \code{rma.uni} from the
\code{metafor} package.}

\item{...}{additional arguments}
}
\value{
a \code{data.frame}
}
\description{
Augment the original data with model residuals, fitted values, and influence
statistics.
}
\examples{

library(broom)
library(metafor)
rma(yi = lnes, sei = selnes, slab = study_name, data = iud_cxca) \%>\%
  augment()

}
