% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{cumulative}
\alias{cumulative}
\title{Title}
\usage{
cumulative(x, prefix = "cumul_", conf.int = FALSE, exponentiate = FALSE,
  glance = FALSE, .f = metafor::cumul, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
