% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_funnel.R
\name{geom_funnel}
\alias{geom_funnel}
\title{Funnel plot lines}
\usage{
geom_funnel(mapping = NULL, data = NULL, ci_level = 0.95,
  log_summary = FALSE, colour = "black", color = NULL,
  position = "identity", na.rm = FALSE,
  summary_label = "Summary\\nEstimate", ci_label = NULL, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{ci_level}{the confidence level desired for the psuedo-CI lines. Between
0 and 1. Also used for line labels.}

\item{log_summary}{a summary estimate (see \code{\link[=pull_summary]{pull_summary()}})}

\item{color, colour}{line color}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{summary_label}{a string. The legend label for the summary line.}

\item{ci_label}{a string. The legend label for the CI lines. Default is
\code{NULL}, in which case the labels will be created based on \code{ci_level}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
`geom_funnel()`` computes psuedo confidence limits for the summary estimate
of a meta-analysis to be used in a funnel plot.
}
\examples{

library(ggplot2)

ma <- meta_analysis(iud_cxca, yi = lnes, sei = selnes, slab = study_name)

ggplot(ma, aes(x = estimate, y = std.error)) +
  geom_point() +
  geom_funnel(aes(summary = pull_summary(ma))) +
  scale_y_reverse()

}
