% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance.rma.uni}
\alias{glance.rma.uni}
\title{Construct a one-row summary of meta-analysis model fit statistics.}
\usage{
glance.rma.uni(x, ...)
}
\arguments{
\item{x}{a meta-analysis object. Currently supports \code{rma.uni} from the
\code{metafor} package.}

\item{...}{additional arguments}
}
\value{
a \code{data.frame}
}
\description{
\code{glance()} computes a one-row summary of  meta-analysis objects,
including estimates of heterogenity and model fit.
}
\examples{

library(broom)
library(metafor)
rma(yi = lnes, sei = selnes, slab = study_name, data = iud_cxca) \%>\%
  glance()

}
