% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iud_cxca}
\alias{iud_cxca}
\title{Meta-analysis of IUD and cervical cancer}
\format{A \code{tbl} with variables about both study esimates and study methods,
including variables controlled, study design, and publication year.}
\source{
Cortessis, V. K., Barrett, M., Brown Wade, N., Enebish, T., Perrigo,
J., Tobin, J., Zhong, C., Zink, J.,Isiaka, V., Muderspach, L., Natavio, M.,
& McKean-Cowdin, R. (2017). “Intrauterine device use and risk of cervical
cancer: a systematic review and meta-analysis.” Obstetrics and Gynecology:
130 (6).
}
\usage{
iud_cxca
}
\description{
This is data from a meta-analysis of the effect of IUD use on cervical
cancer. Estimates are odds ratios for ever-use of IUD.
}
\keyword{datasets}
