% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_funnel.R
\name{scale_x_log}
\alias{scale_x_log}
\alias{scale_y_log}
\title{Log scales for continuous data (x & y)}
\usage{
scale_x_log(..., breaks = scales::log_breaks())

scale_y_log(..., breaks = scales::log_breaks())
}
\arguments{
\item{...}{additional arguments passed to \code{scale_*_continuous()}}

\item{breaks}{a vector of axis breaks}
}
\description{
Log scales for continuous data (x & y)
}
