% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_meta}
\alias{theme_meta}
\alias{theme_forest}
\title{Minimalist theme with meta-analyses in mind}
\usage{
theme_meta(font_size = 14, font_family = "", line_size = 0.5)

theme_forest(font_size = 14, font_family = "", line_size = 0.5)
}
\arguments{
\item{font_size}{font size}

\item{font_family}{font family}

\item{line_size}{line size}
}
\description{
\code{theme_meta()} is a base theme, while \code{theme_forest()} has forest plots in
mind, meaning it removes most legends and sets squares and diamonds as the
default point shapes.
}
