% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{forest_plot}
\alias{forest_plot}
\alias{influence_plot}
\title{Title}
\usage{
forest_plot(x, estimate = estimate, study = study, size = weight,
  shape = type, col = type, xmin = conf.low, xmax = conf.high,
  group = NULL, alpha = 0.75, height = 0, ...)

influence_plot(x, estimate = l1o_estimate, study = study, size = 4,
  shape = 15, col = type, xmin = l1o_conf.low, xmax = l1o_conf.high,
  group = NULL, alpha = 0.75, height = 0, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{

}
