% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\alias{add_weights}
\alias{add_residuals}
\alias{add_rstandard}
\alias{add_i2}
\alias{add_h2}
\alias{add_tau2}
\alias{add_blups}
\title{Title}
\usage{
add_weights(x, .meta = NULL, group = NULL)

add_residuals(x, .meta = NULL)

add_rstandard(x, .meta = NULL)

add_i2(x, .meta = NULL)

add_h2(x, .meta = NULL)

add_tau2(x, .meta = NULL)

add_blups(x, .meta = NULL, exponentiate = FALSE)
}
\arguments{
\item{.meta}{}
}
\description{
Title

#' @export
#' @rdname add
add_predictions <- function(x, .meta = NULL) {
pull_meta(x) %>%
predict() %>%
tibble::enframe("study", "weight") %>%
dplyr::left_join(x, ., by = "study")
}
}
\examples{

}
