% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{add_meta}
\alias{add_meta}
\title{Add a meta-analysis to a tbl}
\usage{
add_meta(x, meta, ...)
}
\arguments{
\item{x}{a \code{data.frame} produced by \code{tidy()}}

\item{meta}{a meta-analysis object (e.g. \code{rma.uni})}

\item{...}{additional arguments to \code{as.tbl}}
}
\value{
a \code{tbl}
}
\description{
Add a meta-analysis to a tbl
}
