% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{meta_analysis}
\alias{meta_analysis}
\title{Meta-Analysis}
\usage{
meta_analysis(.data, .f = metafor::rma, ..., conf.int = TRUE,
  exponentiate = FALSE, include_studies = TRUE, bind_data = FALSE,
  weights = TRUE)
}
\arguments{
\item{.data}{a \code{data.frame} containing the data for the meta-analysis}

\item{.f}{a meta-analysis function. Default is \link[metafor:rma]{metafor::rma}}

\item{...}{additional arguments passed to \code{.f}}

\item{conf.int}{logical. Include confidence intervals?}

\item{exponentiate}{logical. Should the estimates and (if \code{conf.int} =
\code{TRUE}) confidence intervals be exponentiated?}
}
\value{
a \code{tbl}
}
\description{
A wrapper for meta-analytic functions that automatically tidies it and adds
the meta-analysis object to the \code{tbl} in a \code{listcol}.
}
\examples{

}
