% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{sensitivity}
\alias{sensitivity}
\title{Title}
\usage{
sensitivity(x, type = "leave1out", prefix = "l1o_", conf.int = TRUE,
  exponentiate = FALSE, glance = FALSE, .f = metafor::leave1out, ...)
}
\arguments{
\item{x}{a \code{tbl} produced by \code{\link[=meta_analysis]{meta_analysis()}} or by \code{\link[broom:tidy]{broom::tidy()}}}

\item{type}{type of sensitvity analysis.}

\item{...}{}
}
\description{
Title
}
