% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_table.R
\name{text_table}
\alias{text_table}
\title{Title}
\usage{
text_table(.data, study = study, ..., group = NULL, show_y_facets = TRUE,
  show_y_axis = TRUE, show_grid = FALSE, size = 3.75)
}
\arguments{
\item{facets}{}
}
\description{
Title
}
